requirements.txt
folium==0.14.0
haversine==2.8.0
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
plotly==5.9.0
streamlit==1.24.0
streamlit_folium==0.12.0

